+++ doc/zap_resource.lua
@@ -1,6 +1,14 @@
 -- zap_resource
--- @short: Unlink a file in the applpath.
--- @inargs: resstr
+-- @short: Unlink a file in the appltemp path or writable user namespace.
+-- @inargs: string:res
+-- @inargs: string:res, string:namespace
+-- @outargs: bool:success
+-- @longdescr:
+-- This function can be used to delete a file. By default this will apply only
+-- to the appl-temp namespace. If *namespace* is provided and mathces a user-
+-- defined writable namespace ID (see ref:list_namespaces), the deletion will
+-- apply to a file in that namespace instead. Returns whether the file was
+-- successfully deleted or not.
 -- @outargs: boolres
 -- @group: resource
 -- @cfunction: zapresource
@@ -8,7 +16,7 @@
 function main()
 #ifdef MAIN
 	open_rawresource("test.out");
-	zap_rawresource("test.out");
+	zap_resource("test.out");
 
 	if (resource("test.out")) then
 		warning("something went horribly wrong.");
@@ -16,10 +24,10 @@ function main()
 #endif
 
 #ifdef ERROR
-	zap_rawresource("/../../../../../../usr/bin/arcan");
+	zap_resource("/../../../../../../usr/bin/arcan");
 #endif
 
 #ifdef ERROR2
-	zap_rawresource("/../resources/scripts/mouse.lua");
+	zap_resource("/../resources/scripts/mouse.lua");
 #endif
 end