+++ tests/interactive/namespace/namespace.lua
@@ -1,0 +1,7 @@
+-- testing user namespaces
+function namespace()
+	local list = glob_resource(":namespaces")
+	for i, v in ipairs(list) do
+		print(i, v)
+	end
+end